WIP: Whole subsystem test.
